优化 ascii_to_wstr()、wstr_to_ascii()、utf8_to_wstr()、wstr_to_utf8()

+++ src/nut/numeric/numeric_algo/mod.hpp
@@ -479,7 +479,7 @@ inline BigInteger _odd_mod_pow(const BigInteger& a, const BigInteger& b, const B
             ++squre_count;
             bool term = true;
             for (size_t i = 1;
-                 squre_count + i <= wnd_size && bits_left >= i;
+                 squre_count + i <= wnd_size && bits_left >= (int) i;
                 ++i)
             {
                 if (0 != b.bit_at(bits_left - i))
